refactor(core): de-angularize WhatsNewReader
